# Replace pkg_resources with importlib_metadata