--- a/requirements/base.txt
+++ b/requirements/base.txt
@@ -1,3 +1,4 @@
+importlib-metadata
 cwlgen>=0.4.0
 illusional.wdlgen>=0.2.7
 networkx>=2.1