--- a/setup.py
+++ b/setup.py
@@ -51,6 +51,7 @@ setup(
     packages=["janis_core"]
     + ["janis_core." + p for p in sorted(find_packages("./janis_core"))],
     install_requires=[
+        "importlib-metadata",
         "cwlgen >= 0.4.0",
         "illusional.wdlgen >= 0.2.7",
         "ruamel.yaml",