Hopefully this will address some slowness from running Janis on the cluster. Based on information found:

- https://coldfix.de/2017/02/25/slow-entrypoints/
- https://github.com/pypa/setuptools/issues/510

Similar fixes to address the other corresponding janis packages.