@illusional 2020-01-22
Pre changes:

```
$ time janis -v
--------------------  ------
janis-core            v0.9.2
janis-assistant       v0.9.4
janis-unix            v0.9.0
janis-templates       v0.9.1
janis-pipelines       v0.9.1
janis-bioinformatics  v0.9.2
--------------------  ------

real	0m1.065s
user	0m0.586s
sys	0m0.216s
```