@illusional 2020-01-23
Stats from cluster

  | Before | After | Speedup
-- | -- | -- | --
Real | 17.424 | 12.951 |  
  | 14.314 | 12.913 |  
  | 14.335 | 12.574 |  
  | 13.046 | 12.548 |  
  | 13.343 | 13.3 |  
Av | 14.4924 | 12.8572 | 13%
  |   |   |  
User | 1.393 | 1.143 |  
  | 1.244 | 0.994 |  
  | 1.378 | 0.952 |  
  | 1.129 | 0.901 |  
  | 1.193 | 1.026 |  
Av | 1.2674 | 1.0032 | 26%