@illusional 2020-01-23
Interesting, I didn't do one before the removal, but here's a [tuna]() chart on loading `janis -v` on the cluster filesystem.

25% of the time, loading `pkg_resources` for `sqlitedict`, 10% loading `path`, 10% loading `rumael.yaml`.


![tuna-janis](https://user-images.githubusercontent.com/22381693/72956789-2c6f6080-3df5-11ea-8ac9-1548013348aa.png)